# API_SPEC.yaml
openapi: 3.0.0
info:
  title: ForaGo API
  version: 1.0.0
  description: Backend API for ForaGo bushcraft app - PostgreSQL + FastAPI

servers:
  - url: https://api.forago.app
    description: Production
  - url: http://localhost:8000
    description: Development

paths:
  /api/finds/public:
    get:
      summary: Get all public finds (visible to all users)
      tags: [Finds]
      parameters:
        - name: cluster
          in: query
          schema:
            type: string
          description: Filter by clusterHash (optional)
        - name: category
          in: query
          schema:
            type: string
          description: Filter by categoryPath (optional)
        - name: from_date
          in: query
          schema:
            type: string
            format: date-time
          description: Filter from date (optional)
        - name: to_date
          in: query
          schema:
            type: string
            format: date-time
          description: Filter to date (optional)
      responses:
        '200':
          description: List of public finds
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/PublicFindRecord'

  /api/finds/private:
    get:
      summary: Get current user's private finds
      tags: [Finds]
      security:
        - BearerAuth: []
      responses:
        '200':
          description: List of user's private finds
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/PrivateFindRecord'
        '401':
          description: Unauthorized

  /api/finds:
    post:
      summary: Create new find
      tags: [Finds]
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateFindRequest'
      responses:
        '201':
          description: Find created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PrivateFindRecord'
        '400':
          description: Invalid input
        '401':
          description: Unauthorized

  /api/finds/{id}:
    get:
      summary: Get find by ID (if user has access)
      tags: [Finds]
      security:
        - BearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Find details
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/FindRecord'
        '404':
          description: Find not found
        '401':
          description: Unauthorized

    put:
      summary: Update find
      tags: [Finds]
      security:
        - BearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateFindRequest'
      responses:
        '200':
          description: Find updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PrivateFindRecord'
        '404':
          description: Find not found
        '401':
          description: Unauthorized

    delete:
      summary: Delete find
      tags: [Finds]
      security:
        - BearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        '204':
          description: Find deleted
        '404':
          description: Find not found
        '401':
          description: Unauthorized

  /api/finds/nearby:
    get:
      summary: Get finds near a cluster
      tags: [Finds]
      parameters:
        - name: cluster
          in: query
          required: true
          schema:
            type: string
          description: clusterHash to search
        - name: category
          in: query
          schema:
            type: string
          description: Filter by category (optional)
      responses:
        '200':
          description: Nearby finds
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/PublicFindRecord'

  /api/clusters:
    get:
      summary: Get aggregated cluster data
      tags: [Clusters]
      parameters:
        - name: category
          in: query
          schema:
            type: string
          description: Filter by categoryPath (optional)
        - name: from_date
          in: query
          schema:
            type: string
            format: date-time
          description: Filter from date (optional)
        - name: to_date
          in: query
          schema:
            type: string
            format: date-time
          description: Filter to date (optional)
      responses:
        '200':
          description: List of cluster aggregates
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/PublicClusterRecord'

  /api/health:
    get:
      summary: Health check
      tags: [Health]
      responses:
        '200':
          description: API is healthy
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: string
                    example: "ok"

components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT

  schemas:
    PublicFindRecord:
      type: object
      properties:
        id:
          type: string
          format: uuid
        userId:
          type: string
          format: uuid
        date:
          type: string
          format: date-time
        categoryPaths:
          type: array
          items:
            type: array
            items:
              type: string
          example: [["nature", "forest", "tree"], ["edible"]]
        description:
          type: string
        clusterHash:
          type: string
          example: "51.5_-0.1"
        images:
          type: array
          items:
            $ref: '#/components/schemas/RecordImageRef'
        comments:
          type: array
          items:
            $ref: '#/components/schemas/RecordComment'
      required: [id, userId, date, categoryPaths, description, clusterHash]

    PrivateFindRecord:
      allOf:
        - $ref: '#/components/schemas/PublicFindRecord'
        - type: object
          properties:
            location:
              $ref: '#/components/schemas/LatLng'
          required: [location]

    PublicClusterRecord:
      type: object
      properties:
        clusterHash:
          type: string
          example: "51.5_-0.1"
        categoryPathCounts:
          type: object
          additionalProperties:
            type: integer
          example:
            "nature/forest/tree": 2
            "nature/water/spring": 1
        totalRecords:
          type: integer
          example: 3
        lastUpdated:
          type: string
          format: date-time
      required: [clusterHash, categoryPathCounts, totalRecords, lastUpdated]

    RecordImageRef:
      type: object
      properties:
        thumbnailUrl:
          type: string
          format: uri
        fullUrl:
          type: string
          format: uri
        storageRef:
          type: string
          nullable: true
          description: Reference in Supabase Storage
      required: [thumbnailUrl, fullUrl]

    RecordComment:
      type: object
      properties:
        id:
          type: string
          format: uuid
        userId:
          type: string
          format: uuid
        text:
          type: string
        createdAt:
          type: string
          format: date-time
      required: [id, userId, text, createdAt]

    LatLng:
      type: object
      properties:
        latitude:
          type: number
          format: double
        longitude:
          type: number
          format: double
      required: [latitude, longitude]

    CreateFindRequest:
      type: object
      properties:
        date:
          type: string
          format: date-time
        categoryPaths:
          type: array
          items:
            type: array
            items:
              type: string
        description:
          type: string
        location:
          $ref: '#/components/schemas/LatLng'
        clusterHash:
          type: string
      required: [date, categoryPaths, description, location, clusterHash]

    UpdateFindRequest:
      type: object
      properties:
        date:
          type: string
          format: date-time
        categoryPaths:
          type: array
          items:
            type: array
            items:
              type: string
        description:
          type: string
        location:
          $ref: '#/components/schemas/LatLng'
      minProperties: 1

    FindRecord:
      oneOf:
        - $ref: '#/components/schemas/PublicFindRecord'
        - $ref: '#/components/schemas/PrivateFindRecord'

tags:
  - name: Finds
    description: Operations for finding records
  - name: Clusters
    description: Aggregated cluster data
  - name: Health
    description: Health check